backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
paho-mqtt==2.1.0
